# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watchlist:
    get:
      tags: [watchlist]
      operationId: getWatchlist
      summary: Get the authenticated user's watchlist
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Watchlist items.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [watchlist]
      operationId: addToWatchlist
      summary: Add a symbol to the watchlist
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WatchlistAddRequest'
      responses:
        '200':
          description: Item added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistItem'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /watchlist/{symbol}:
    delete:
      tags: [watchlist]
      operationId: removeFromWatchlist
      summary: Remove a symbol from the watchlist
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  parameters:
...
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    WatchlistItem:
      type: object
      required: [id, userId, symbol, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        symbol:
          type: string
        note:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    WatchlistResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WatchlistItem'
    WatchlistAddRequest:
      type: object
      required: [symbol]
      properties:
        symbol:
          type: string
          minLength: 1
        note:
          type: string