# lib/api-zod/src/generated/types/watchlistItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WatchlistItem {
  id: number;
  userId: string;
  symbol: string;
  /** @nullable */
  note?: string | null;
  createdAt: Date;
}